+++ src/lib/contexts/my-user-context.svelte.ts
@@ -571,5 +571,5 @@ export const myUserContext = new MyUserContext();
 
 // Expose myUserContext to window for E2E testing
 if (typeof window !== 'undefined') {
-  (window as any).__myUserContext = myUserContext;
+  (window as Window & typeof globalThis & { __myUserContext: typeof myUserContext }).__myUserContext = myUserContext;
 }
